test: rename MAIN import to main_module
